Added support of scalajs features and added two new features for sharing and tags interactivity

+++ models/src/main/scala/com/tkachuko/blog/models/package.scala
@@ -5,7 +5,7 @@ package object models {
   case class Post(title: String,
                   content: String,
                   tags: List[String] = List.empty,
-                  created: Long = System.currentTimeMillis())
+                  created: Double = System.currentTimeMillis())
 
   case class Subscription(email: String)
